BoundedSSA: Distributions as a direct dep, drop the extension
Per review (Rackauckas): the uniformization SSA calls `rand(Bernoulli(p))` directly, so there is no glue left for StochasticAD to hook. Delete ext/JumpProcessesStochasticADExt.jl and the `_bounded_ssa_bernoulli` indirection; StochasticAD differentiates `rand(::Bernoulli)` itself once the user loads it and passes a StochasticTriple parameter. Distributions moves from [weakdeps] to [deps] (+[compat] 0.25); StochasticAD drops out of weakdeps/extensions/compat and is now only a test dpendency.

NOTE for weight/benchmarking review: this makes Distributions a hard dependency of JumpProcesses (it pulls ~20 transitive packages). Reverting to a weakdep + a one-line `_bounded_ssa_bernoulli` hook in a re-created extension is mechanical if the added weight is not acceptable.

Only `BoundedSSA` is public: the struct + docstring are consolidated into src/bounded_ssa.jl, `bounded_ssa_path` is unexported (tests call it qualified), and BoundedSSA is documented in docs/src/api.md and docs/src/jump_solve.md.

Run the StochasticAD test group in a fresh subprocess (test/runtests.jl) so only the isolated `stochasticad` project is loaded, eliminating spurious precompile /
version-skew warnings from having two package environments in one process. A failing @testset exits nonzero, so the suite still fails on a regression.

Verified on Julia 1.12: GROUP=StochasticAD 16/16, GROUP=QA 11/11.
