+++ src/SSA_stepper.jl
@@ -60,78 +60,6 @@ struct SSAStepper <: SciMLBase.AbstractDEAlgorithm end
 SciMLBase.allows_late_binding_tstops(::SSAStepper) = true
 
 """
-    BoundedSSA(; rate_bound)
-
-A StochasticAD-compatible SSA algorithm for **jump-only** `ConstantRateJump`
-`DiscreteProblem`s, giving correct gradients via StochasticAD's
-`derivative_estimate`/`stochastic_triple` — with `saveat` support, so the whole
-sampled path is differentiable, not only the terminal state.
-
-The stock `SSAStepper` cannot be differentiated with StochasticAD: it advances
-time with a `while integrator.t < integrator.tstop < end_time` loop, i.e. a
-boolean predicate on (triple-valued) time, which StochasticAD forbids by design —
-so the event-count derivative is dropped (a state-dependent rate yields a gradient
-of `0`). `BoundedSSA` instead uses **uniformization (thinning)** against a fixed
-total-propensity bound `Λ = rate_bound`:
-
-  - candidate event times form a homogeneous Poisson process of rate `Λ` on the
-    time span — these are **parameter-free**, so the loop never branches on a
-    triple and the times stay `Float64`;
-  - at each candidate the current total propensity `a(u)` is recomputed and the
-    event is *accepted* with a tracked `Bernoulli(a(u)/Λ)` (otherwise it is a
-    **null event** absorbing the slack `Λ - a(u)`);
-  - the firing channel is chosen by stick-breaking `Bernoulli`s.
-
-All parameter dependence flows through the accept / channel `Bernoulli`s, so the
-gradient is captured. This is **unbiased** (no step cap) whenever `Λ` is a valid
-bound, and `saveat` is exact because the candidate times are fixed `Float64`.
-
-With ordinary `Float64` parameters `solve(jprob, BoundedSSA(; rate_bound))` is an
-ordinary (uniformization) SSA simulation; with StochasticAD triples it
-differentiates.
-
-# Keyword arguments
-
-  - `rate_bound` (required): a constant `Λ` upper-bounding the **total** propensity
-    `Σₖ rateₖ(u, p, t)` over the whole trajectory (and over the parameter
-    perturbation). Valid for systems with rigorously bounded populations; a looser
-    bound only costs efficiency (more null events), not accuracy. If `Λ` is
-    violated the accept probability exceeds 1 and sampling errors — pick it with
-    margin.
-
-# `solve` options
-
-  - `saveat`: times (a vector, or a `Number` step) at which to return the solution,
-    with `save_start`/`save_end` controlling the endpoints (same conventions as
-    `SimpleTauLeaping`, via `_process_saveat`); defaults to `[t0, tf]`. `sol.u[i]` is
-    the differentiable state at `sol.t[i]`, and `sol(t)` interpolates (piecewise
-    constant, as with `SSAStepper`).
-
-# Scope / limitations
-
-  - `ConstantRateJump`s only (state-dependent rates supported); jump-only, no
-    continuous drift, no `VariableRateJump`. `MassActionJump` is not yet supported.
-  - Additive affects only (the net change is inferred from `affect!` and checked).
-  - The differentiation parameter `prob.p` must be a summable/indexable numeric
-    collection (e.g. a `Vector`). SciMLStructures parameter objects (MTK/Catalyst
-    tunables) are not yet specialized — a documented follow-up.
-  - The solver itself is plain (no StochasticAD dependency): with ordinary parameters
-    `solve(jprob, BoundedSSA(; rate_bound))` is a uniformization SSA simulation. It
-    becomes differentiable only when `StochasticAD` (and `Distributions`) are loaded,
-    which supplies the differentiable accept/channel decision.
-
-See also [`bounded_ssa_path`](@ref), the differentiable core this wraps.
-"""
-struct BoundedSSA{B} <: DiffEqBase.AbstractDEAlgorithm
-    rate_bound::B
-end
-function BoundedSSA(; rate_bound = nothing)
-    rate_bound === nothing && error("BoundedSSA requires the keyword argument " *
-        "`rate_bound` (a constant upper bound on the total propensity).")
-    BoundedSSA{typeof(rate_bound)}(rate_bound)
-end
-
-"""
 $(TYPEDEF)
 
 Solution objects for pure jump problems solved via `SSAStepper`.
